use smaller sized elements in attempt to reduce memory footprint

+++ SortItem.h
@@ -1,12 +1,14 @@
 #ifndef SORTITEM_H
 #define SORTITEM_H
 
+#include <QtGlobal>
+
 class SortItem
 {
 public:
 	bool operator <(const SortItem &other) const;
 	float mDistance;
-	int mClass;
+    qint8 mClass;
 };
 
 #endif // SORTITEM_H
